+++ LookupAnything/Framework/TargetFactory.cs
@@ -51,7 +51,7 @@ namespace Pathoschild.Stardew.LookupAnything.Framework
                 new BuildingLookupProvider(reflection, gameHelper, codex),
                 new CharacterLookupProvider(reflection, gameHelper, config, codex),
                 new ItemLookupProvider(reflection, gameHelper, config, codex, jsonAssets),
-                new TerrainFeatureLookupProvider(reflection, gameHelper, jsonAssets),
+                new TerrainFeatureLookupProvider(reflection, gameHelper, codex, jsonAssets),
                 new TileLookupProvider(reflection, gameHelper, config, showRawTileInfo)
             };
         }